requirements.txt
# Core Packages
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.0
scikit-learn==1.3.2
Pillow==10.1.0

# Optional Visualization Packages
plotly==5.18.0
matplotlib==3.8.2

# Development Tools (Optional)
pytest==7.4.3
black==23.11.0